Installed Grunt
Grunt itself can be installed locally. In fact, it should
be installed locally because other projects might have dependency
with a different Grunt version.

On the other hand, the Grunt Client (grunt-cli) should be installed
globally since it is just an interface for Grunt.

The directory node_modules is ignored on git repository.

Changes to be committed:
	modified:   .gitignore
	modified:   bootstrap.sh
	modified:   public/package.json